Update remaining dependencies and trim direct deps
Move Flux from a direct dependency to a weak dependency with a new
TopOptFluxExt package extension. The NeuralNetwork struct and all
non-Flux methods stay in src/Functions/neural.jl; only the two
constructors that call Flux.destructure move to the extension.

Remove Revise and JuliaFormatter from direct deps (dev-only tools).
Upgrade JSON 0.21 -> 1.7.0 and StatsFuns 1.5 -> 2.2.1. Widen compat
bounds for ForwardDiff (0.10, 1), TimerOutputs (0.5, 1), and Flux
(0.14, 0.15, 0.16) so they upgrade automatically once transitive
blockers (ADNLPModels, LinearOperators, NonconvexMMA) are resolved.

Fixes #214
