+++ Project.toml
@@ -11,13 +11,12 @@ Distributions = "0.24, 0.25"
 Einsum = "0.4"
 Ferrite = "1"
 FileIO = "1"
-Flux = "0.11, 0.12, 0.13, 0.14"
-ForwardDiff = "<0.10.35, 0.10, 1"
+Flux = "0.14, 0.15, 0.16"
+ForwardDiff = "0.10, 1"
 GeometryBasics = "0.5"
 GeometryTypes = "0.8"
 IterativeSolvers = "0.9"
-JSON = "0.21"
-JuliaFormatter = "1"
+JSON = "0.21, 1"
 MacroTools = "0.5"
 Makie = "0.23"
 MappedArrays = "0.4"
@@ -30,13 +29,12 @@ Parameters = "0.12, 0.13"
 Preconditioners = "0.3, 0.4, 0.5, 0.6"
 Preferences = "1"
 Reexport = "1"
-Revise = "3"
 Setfield = "0.7, 0.8, 1"
 StaticArrays = "1"
 Statistics = "1"
-StatsFuns = "0.9, 1"
+StatsFuns = "0.9, 1, 2"
 Tensors = "1.4"
-TimerOutputs = "0.5"
+TimerOutputs = "0.5, 1"
 VTKDataTypes = "0.1"
 WriteVTK = "1.9"
 Zygote = "0.6"
@@ -50,13 +48,11 @@ Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Einsum = "b7d42ee7-0b51-5a75-98ca-779d3107e4c0"
 Ferrite = "c061ca5d-56c9-439f-9c0e-210fe06d3992"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
-Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
 GeometryTypes = "4d00f742-c7ba-57c2-abde-4428a4b178cb"
 IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
-JuliaFormatter = "98e50ef6-434e-11e9-1051-2b60c6c9e899"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 MappedArrays = "dbb5928d-eab1-5f90-85c2-b9b0edb7c900"
@@ -71,7 +67,6 @@ Preconditioners = "af69fa37-3177-5a40-98ee-561f696e4fcd"
 Preferences = "21216c6a-2e73-6563-6e65-726566657250"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 Setfield = "efcf1570-3423-57d1-acb7-fd33fddbac46"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
@@ -86,6 +81,8 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [extensions]
 TopOptMakieExt = "Makie"
+TopOptFluxExt = "Flux"
 
 [weakdeps]
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
